requirements.txt
arrow==1.3.0
asgiref==3.7.2
attrs==23.2.0
click==8.1.7
comm==0.2.2
crispy-tailwind==1.0.3
debugpy==1.8.1
decorator==5.1.1
Django==5.0.1
django-appconf==1.0.6
django-crispy-forms==2.1
django-environ==0.11.2
django-select2==8.1.2
django-tailwind==3.8.0
gunicorn==21.2.0
ipykernel==6.29.3
ipython==8.22.2
Jinja2==3.1.3
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
pandas==2.1.4
pillow==10.2.0
pipenv==2023.12.1
psycopg2-binary==2.9.9
Pygments==2.17.2
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.1.0
six==1.16.0
sqids==0.4.1
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.20240106
tzdata==2023.4
whitenoise==6.6.0
